+++ slackdown.go
@@ -2,21 +2,12 @@ package slackdown
 
 import (
 	"bytes"
+	"fmt"
 	"io"
 	"strconv"
 
 	bf "github.com/russross/blackfriday/v2"
 )
-
-// Renderer is the rendering interface for slack output.
-type Renderer struct {
-	w             bytes.Buffer
-	lastOutputLen int
-}
-
-var itemLevel = 0
-
-var itemListMap = make(map[int]int)
 
 var (
 	strongTag        = []byte("*")
@@ -27,38 +18,86 @@ var (
 	linkTag          = []byte("<")
 	linkCloseTag     = []byte(">")
 	pipeSign         = []byte("|")
-)
 
-var (
 	nlBytes    = []byte{'\n'}
 	spaceBytes = []byte{' '}
+
+	escapes = [256][]byte{
+		'&': []byte(`&amp;`),
+		'<': []byte(`&lt;`),
+		'>': []byte(`&gt;`),
+	}
 )
 
-var escapes = [256][]byte{
-	'&': []byte(`&amp;`),
-	'<': []byte(`&lt;`),
-	'>': []byte(`&gt;`),
+// Renderer is the rendering interface for slack output.
+type Renderer struct {
+	lastOutputLen int
+	itemLevel     int
+	itemListDepth map[int]int
+	err           error
+}
+
+func NewRenderer() *Renderer {
+	return &Renderer{
+		itemListDepth: make(map[int]int),
+	}
+}
+
+// Err will return a non-nil error if Render methods failed, usually
+// from an io.Writer error. It should be called after a Render call
+// but before a call to Reset.
+func (r *Renderer) Err() error { return r.err }
+
+// Reset optionally resets the internal state of the Renderer for reuse.
+func (r *Renderer) Reset() {
+	r.lastOutputLen = 0
+	r.itemLevel = 0
+	r.err = nil
+
+	for k := range r.itemListDepth {
+		delete(r.itemListDepth, k)
+	}
 }
 
 func (r *Renderer) esc(w io.Writer, text []byte) {
+	if r.err != nil {
+		return
+	}
+
 	var start, end int
-	for end < len(text) {
-		if escSeq := escapes[text[end]]; escSeq != nil {
-			w.Write(text[start:end])
-			w.Write(escSeq)
-			start = end + 1
+	for ; end < len(text); end++ {
+		escSeq := escapes[text[end]]
+		if escSeq == nil {
+			continue
 		}
-		end++
+
+		_, e1 := w.Write(text[start:end])
+		_, r.err = w.Write(escSeq)
+		start = end + 1
+
+		if e1 != nil {
+			r.err = e1
+		}
+
+		if r.err != nil {
+			return
+		}
 	}
 
 	if start < len(text) && end <= len(text) {
-		w.Write(text[start:end])
+		_, r.err = w.Write(text[start:end])
 	}
 }
 
 func (r *Renderer) out(w io.Writer, text []byte) {
-	w.Write(text)
-	r.lastOutputLen = len(text)
+	if r.err != nil {
+		return
+	}
+
+	n, err := w.Write(text)
+	r.lastOutputLen = n
+
+	r.err = err
 }
 
 func (r *Renderer) cr(w io.Writer) {
@@ -74,25 +113,25 @@ func (r *Renderer) RenderNode(w io.Writer, node *bf.Node, entering bool) bf.Walk
 	case bf.Text:
 		r.esc(w, node.Literal)
 	case bf.Softbreak:
-		break
+
 	case bf.Hardbreak:
-		break
+
 	case bf.BlockQuote:
-		break
+
 	case bf.CodeBlock:
 		r.out(w, codeBlockTag)
 		r.esc(w, node.Literal)
 		r.out(w, codeBlockTag)
 		r.cr(w)
 		r.cr(w)
-		break
+
 	case bf.Code:
 		r.out(w, codeTag)
 		r.esc(w, node.Literal)
 		r.out(w, codeTag)
-		break
+
 	case bf.Emph:
-		break
+
 	case bf.Heading:
 		if entering {
 			r.out(w, strongTag)
@@ -101,24 +140,24 @@ func (r *Renderer) RenderNode(w io.Writer, node *bf.Node, entering bool) bf.Walk
 			r.cr(w)
 		}
 	case bf.Image:
-		break
+
 	case bf.Item:
 		if entering {
 			r.out(w, spaceBytes)
-			for i := 1; i < itemLevel; i++ {
+			for i := 1; i < r.itemLevel; i++ {
 				r.out(w, spaceBytes)
 				r.out(w, spaceBytes)
 				r.out(w, spaceBytes)
 			}
 			if node.ListFlags&bf.ListTypeOrdered != 0 {
-				r.out(w, append([]byte(strconv.Itoa(itemListMap[itemLevel])), node.ListData.Delimiter))
-				itemListMap[itemLevel]++
+				r.out(w, append([]byte(strconv.Itoa(r.itemListDepth[r.itemLevel])), node.ListData.Delimiter))
+				r.itemListDepth[r.itemLevel]++
 			} else {
 				r.out(w, itemTag)
 			}
 			r.out(w, spaceBytes)
 		}
-		break
+
 	case bf.Link:
 		if entering {
 			r.out(w, linkTag)
@@ -129,27 +168,27 @@ func (r *Renderer) RenderNode(w io.Writer, node *bf.Node, entering bool) bf.Walk
 		} else {
 			r.out(w, linkCloseTag)
 		}
-		break
+
 	case bf.HorizontalRule:
-		break
+
 	case bf.List:
 		if entering {
-			itemLevel++
+			r.itemLevel++
 			if node.ListFlags&bf.ListTypeOrdered != 0 {
-				itemListMap[itemLevel] = 1
+				r.itemListDepth[r.itemLevel] = 1
 			}
 		} else {
 			if node.ListFlags&bf.ListTypeOrdered != 0 {
-				delete(itemListMap, itemLevel)
+				delete(r.itemListDepth, r.itemLevel)
 			}
-			itemLevel--
-			if itemLevel == 0 {
+			r.itemLevel--
+			if r.itemLevel == 0 {
 				r.cr(w)
 			}
 		}
-		break
+
 	case bf.Document:
-		break
+
 	case bf.Paragraph:
 		if !entering {
 			if node.Parent.Type != bf.Item {
@@ -157,40 +196,59 @@ func (r *Renderer) RenderNode(w io.Writer, node *bf.Node, entering bool) bf.Walk
 			}
 			r.cr(w)
 		}
-		break
+
 	case bf.Strong:
 		r.out(w, strongTag)
-		break
+
 	case bf.Del:
 		r.out(w, strikethroughTag)
-		break
+
 	case bf.HTMLBlock:
-		break
+
 	case bf.HTMLSpan:
-		break
+
 	case bf.Table:
-		break
+
 	case bf.TableCell:
-		break
+
 	case bf.TableHead:
-		break
+
 	case bf.TableBody:
-		break
+
 	case bf.TableRow:
-		break
+
 	default:
-		panic("Unknown node type " + node.Type.String())
+		if r.err == nil {
+			r.err = fmt.Errorf("unknown node type: %s", node.Type)
+		}
 	}
+
+	if r.err != nil {
+		return bf.Terminate
+	}
+
 	return bf.GoToNext
 }
 
 // Render prints out the whole document from the ast.
+// A call to Err will return non-nil if rendering stopped early,
+// usually due to an internal writting error.
 func (r *Renderer) Render(ast *bf.Node) []byte {
+	buf := new(bytes.Buffer)
+	r.RenderOut(buf, ast)
+
+	return buf.Bytes()
+}
+
+// RenderOut renders the ast to the provided io.Writer. An error
+// is returned if rendering stopped early, usually due to an io.Writer error.
+// The error returned from this method is the same as a call to Err.
+func (r *Renderer) RenderOut(w io.Writer, ast *bf.Node) error {
 	ast.Walk(func(node *bf.Node, entering bool) bf.WalkStatus {
-		return r.RenderNode(&r.w, node, entering)
+		return r.RenderNode(w, node, entering)
 	})
 
-	return r.w.Bytes()
+	return r.Err()
 }
 
 // RenderHeader writes document header (unused).
@@ -203,7 +261,7 @@ func (r *Renderer) RenderFooter(w io.Writer, ast *bf.Node) {
 
 // Run prints out the confluence document.
 func Run(input []byte, opts ...bf.Option) []byte {
-	r := &Renderer{}
+	r := NewRenderer()
 	optList := []bf.Option{bf.WithRenderer(r), bf.WithExtensions(bf.CommonExtensions)}
 	optList = append(optList, opts...)
 	parser := bf.New(optList...)